+++ tools/qfcc/test/ptrfunc.r
@@ -28,7 +28,7 @@ find_xdef (void *varaddr)
 	//FIXME need a simple way to get at a def's meta-data
 	xdefs_t    *xdefs = PR_FindGlobal (".xdefs");
 	xdef_t     *xdef = xdefs.xdefs;
-	while (xdef - xdefs.xdefs < xdefs.num_xdefs && xdef.ofs != varaddr) {
+	while (xdef - xdefs.xdefs < (int) xdefs.num_xdefs && xdef.ofs != varaddr) {
 		xdef++;
 	}
 	if (xdef.ofs != varaddr) {
